+++ src/shared/components/common/QRCode/QRCode.tsx
@@ -1,5 +1,6 @@
 import { QRCodeSVG } from 'qrcode.react';
 import { Center, Paper } from '../Layout';
+import { AspectRatio } from '../Layout/AspectRatio';
 
 type Props = {
   url: string;
@@ -11,23 +12,25 @@ export const QRCode = (props: Props) => {
   return (
     <Center>
       <Paper radius="md" withBorder p="md" pb={10}>
-        <QRCodeSVG
-          value={url}
-          size={size}
-          bgColor={'#ffffff'}
-          fgColor={'#000000'}
-          level={'L'}
-          {...(imageSrc && {
-            imageSettings: {
-              src: imageSrc,
-              x: undefined,
-              y: undefined,
-              height: size / 4.5,
-              width: size / 4.5,
-              excavate: true,
-            },
-          })}
-        />
+        <AspectRatio ratio={1} w={size} h={size}>
+          <QRCodeSVG
+            value={url}
+            // size={size}
+            bgColor={'#ffffff'}
+            fgColor={'#000000'}
+            level={'L'}
+            {...(imageSrc && {
+              imageSettings: {
+                src: imageSrc,
+                x: undefined,
+                y: undefined,
+                height: size / 4.5,
+                width: size / 4.5,
+                excavate: true,
+              },
+            })}
+          />
+        </AspectRatio>
       </Paper>
     </Center>
   );
